remove session from property method arguments, other misc review changes
