+++ tests/workflows/test_ophys_experiment.py
@@ -11,10 +11,9 @@ setup_app_config(
             Path(__file__).parent / 'resources' / 'config.yml'),
     test_di_base_model_path=Path(__file__).parent / 'resources' / 'di_model.h5'
 )
-
+from unittest.mock import patch
 from ophys_etl.workflows.db.initialize_db import InitializeDBRunner # noqa E402
 from sqlmodel import Session
-from ophys_etl.test_utils.db_base import MockSQLiteDB
 from ophys_etl.workflows.ophys_experiment import OphysExperiment, \
     OphysSession, Specimen
 from ophys_etl.workflows.workflow_names import WorkflowNameEnum
@@ -22,7 +21,6 @@ from ophys_etl.workflows.workflow_steps import WorkflowStepEnum
 from ophys_etl.workflows.workflow_step_runs import get_workflow_step_by_name
 from ophys_etl.workflows.db.schemas import WorkflowStepRun, MotionCorrectionRun, \
     OphysROI, OphysROIMaskValue
-
 
 class TestOphysExperiment:
     @classmethod
@@ -97,39 +95,43 @@ class TestOphysExperiment:
     def setup(self):
         self._initializeDB()
         self._create_mock_data()
-        self.ophys_experiment = OphysExperiment(
-                    id='1',
-                    session=OphysSession(id='2'),
-                    specimen=Specimen(id='3'),
-                    storage_directory=Path('/storage_dir'),
-                    raw_movie_filename=Path('mov.h5'),
-                    movie_frame_rate_hz=11.0
-                )
 
     def teardown_method(cls):
         shutil.rmtree(cls._tmp_dir)
 
 
     def test__get_ophys_experiment_roi_metadata(self):
-        with Session(self._engine) as session:
-            roi_metadata = self.ophys_experiment.get_ophys_experiment_roi_metadata(
-                session=session,
+        with patch('ophys_etl.workflows.db.engine', self._engine):
+            ophys_experiment = OphysExperiment(
+                    id='1',
+                    session=OphysSession(id='2'),
+                    specimen=Specimen(id='3'),
+                    storage_directory=Path('/storage_dir'),
+                    raw_movie_filename=Path('mov.h5'),
+                    movie_frame_rate_hz=11.0
             )
-        assert len(roi_metadata) == 1
-        assert roi_metadata[0]["x"] == 10
-        assert roi_metadata[0]["y"] == 20
-        assert roi_metadata[0]["width"] == 30
-        assert roi_metadata[0]["height"] == 40
-        assert roi_metadata[0]["mask"][5][6] == 1
-        assert len(roi_metadata[0]["mask"]) == 40
-        assert len(roi_metadata[0]["mask"][0]) == 30
+            roi_metadata = ophys_experiment.get_ophys_experiment_roi_metadata
+            assert len(roi_metadata) == 1
+            assert roi_metadata[0]["x"] == 10
+            assert roi_metadata[0]["y"] == 20
+            assert roi_metadata[0]["width"] == 30
+            assert roi_metadata[0]["height"] == 40
+            assert roi_metadata[0]["mask"][5][6] == 1
+            assert len(roi_metadata[0]["mask"]) == 40
+            assert len(roi_metadata[0]["mask"][0]) == 30
     
     def test_get_ophys_experiment_motion_border(self):
-        with Session(self._engine) as session:
-            motion_border = self.ophys_experiment.get_ophys_experiment_motion_border(
-                session=session
+        with patch('ophys_etl.workflows.db.engine', self._engine):
+            ophys_experiment = OphysExperiment(
+                    id='1',
+                    session=OphysSession(id='2'),
+                    specimen=Specimen(id='3'),
+                    storage_directory=Path('/storage_dir'),
+                    raw_movie_filename=Path('mov.h5'),
+                    movie_frame_rate_hz=11.0
             )
-        assert motion_border['x0'] == 10
-        assert motion_border['x1'] == 20
-        assert motion_border['y0'] == 30
-        assert motion_border['y1'] == 40
+            motion_border = ophys_experiment.get_ophys_experiment_motion_border
+            assert motion_border['x0'] == 10
+            assert motion_border['x1'] == 20
+            assert motion_border['y0'] == 30
+            assert motion_border['y1'] == 40
